# backend/docs/openapi.yaml
openapi: 3.1.0
info:
  title: AgroJusto API
  version: 1.0.0
  description: API do AgroJusto para produtores, custos, precificacao e aluguel de maquinas.
servers:
  - url: http://localhost:3333
    description: Desenvolvimento local
security:
  - bearerAuth: []
tags:
  - name: Health
  - name: Auth
  - name: Perfil
  - name: Produtos
  - name: Custos
  - name: Precificacao
  - name: Maquinas
  - name: Alugueis
  - name: Dashboard
paths:
  /health:
    get:
      tags: [Health]
      summary: Verifica a disponibilidade da API
      security: []
      responses:
        '200':
          description: API disponivel
          content:
            application/json:
              schema:
                type: object
                required: [ok, service]
                properties:
                  ok: { type: boolean, example: true }
                  service: { type: string, example: agrojusto-backend }
  /auth/register:
    post:
      tags: [Auth]
      summary: Cria uma conta e inicia uma sessao
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
            example:
              nome: Maria Silva
              email: maria@example.com
              senha: senha-segura-123
              regiao: Sul
      responses:
        '201':
          description: Conta criada
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Session' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '409': { $ref: '#/components/responses/ConflictError' }
  /auth/login:
    post:
      tags: [Auth]
      summary: Inicia uma sessao
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginRequest' }
            example:
              email: maria@example.com
              senha: senha-segura-123
      responses:
        '200':
          description: Login realizado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Session' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/UnauthorizedError' }
  /auth/logout:
    post:
      tags: [Auth]
      summary: Revoga a sessao atual
      responses:
        '200':
          description: Sessao encerrada
          content:
            application/json:
              schema:
                type: object
                properties: { success: { type: boolean, example: true } }
        '401': { $ref: '#/components/responses/UnauthorizedError' }
  /me:
    get:
      tags: [Perfil]
      summary: Retorna o perfil autenticado
      responses:
        '200':
          description: Perfil do usuario
          content:
            application/json:
              schema:
                type: object
                properties: { user: { $ref: '#/components/schemas/User' } }
        '401': { $ref: '#/components/responses/UnauthorizedError' }
    patch:
      tags: [Perfil]
      summary: Atualiza o nome ou regiao do perfil
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ProfileUpdateRequest' }
            example: { nome: Maria Santos, regiao: Sudeste }
      responses:
        '200':
          description: Perfil atualizado
          content:
            application/json:
              schema:
                type: object
                properties: { user: { $ref: '#/components/schemas/User' } }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/UnauthorizedError' }
  /products:
    get:
      tags: [Produtos]
      summary: Lista produtos do usuario
      responses:
        '200':
          description: Lista de produtos
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Product' }
        '401': { $ref: '#/components/responses/UnauthorizedError' }
    post:
      tags: [Produtos]
      summary: Cria um produto
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ProductInput' }
            example:
              nome: Milho
              quantidade: 12.5
              unidade: saca
              cultivo: organico
              regiao: Sul
              preco: 75.5
              imagem: https://example.com/milho.jpg
      responses:
        '201':
          description: Produto criado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Product' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/UnauthorizedError' }
  /products/{id}:
    parameters:
      - $ref: '#/components/parameters/Id'
    patch:
      tags: [Produtos]
      summary: Atualiza um produto
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ProductUpdate' }
      responses:
        '200':
          description: Produto atualizado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Product' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/UnauthorizedError' }
        '404': { $ref: '#/components/responses/NotFoundError' }
    delete:
      tags: [Produtos]
      summary: Exclui um produto
      responses:
        '204': { description: Produto excluido }
        '401': { $ref: '#/components/responses/UnauthorizedError' }
        '404': { $ref: '#/components/responses/NotFoundError' }
  /costs:
    get:
      tags: [Custos]
      summary: Lista custos do usuario
      responses:
        '200':
          description: Lista de custos
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Cost' }
        '401': { $ref: '#/components/responses/UnauthorizedError' }
    post:
      tags: [Custos]
      summary: Cria um custo
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CostInput' }
            example:
              categoria: INSUMO
              descricao: Adubo organico
              valor: 800
              data: '2026-09-10'
      responses:
        '201':
          description: Custo criado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Cost' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/UnauthorizedError' }
        '404': { $ref: '#/components/responses/NotFoundError' }
  /costs/{id}:
    parameters:
      - $ref: '#/components/parameters/Id'
    delete:
      tags: [Custos]
      summary: Exclui um custo
      responses:
        '204': { description: Custo excluido }
        '401': { $ref: '#/components/responses/UnauthorizedError' }
        '404': { $ref: '#/components/responses/NotFoundError' }
  /costs/summary:
    get:
      tags: [Custos]
      summary: Resume custos por categoria
      responses:
        '200':
          description: Resumo de custos
          content:
            application/json:
              schema:
                type: object
                required: [total, categories]
                properties:
                  total: { type: number, format: double, example: 1200.5 }
                  categories:
                    type: array
                    items:
                      type: object
                      required: [categoria, total]
                      properties:
                        categoria: { type: string, example: INSUMO }
                        total: { type: number, format: double, example: 800 }
        '401': { $ref: '#/components/responses/UnauthorizedError' }
  /pricing/simulations:
    post:
      tags: [Precificacao]
      summary: Calcula e salva uma simulacao
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PricingSimulationInput' }
            example:
              insumos: 100
              maoDeObra: 50
              transporteBase: 20
              manutencao: 10
              distanciaKm: 10
              custoPorKm: 2
              quantidade: 10
              margem: 10
              oferta: 1
              demanda: 1.2
      responses:
        '201':
          description: Simulacao criada
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PricingSimulation' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/UnauthorizedError' }
        '404': { $ref: '#/components/responses/NotFoundError' }
    get:
      tags: [Precificacao]
      summary: Lista simulacoes paginadas
      parameters:
        - { name: page, in: query, schema: { type: integer, minimum: 1, default: 1 } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 20 } }
      responses:
        '200':
          description: Historico paginado
          content:
            application/json:
              schema:
                type: object
                properties:
                  items: { type: array, items: { $ref: '#/components/schemas/PricingSimulation' } }
                  page: { type: integer, example: 1 }
                  limit: { type: integer, example: 20 }
                  total: { type: integer, example: 35 }
                  pages: { type: integer, example: 2 }
        '401': { $ref: '#/components/responses/UnauthorizedError' }
  /machines:
    get:
      tags: [Maquinas]
      summary: Lista maquinas ativas do usuario
      responses:
        '200':
          description: Lista de maquinas
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Machine' }
        '401': { $ref: '#/components/responses/UnauthorizedError' }
    post:
      tags: [Maquinas]
      summary: Cadastra uma maquina
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/MachineInput' }
            example:
              nome: Trator
              preco: 250
              tipo: agricola
              condicoes: Boa
              dias: 7
      responses:
        '201':
          description: Maquina criada
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Machine' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/UnauthorizedError' }
  /machines/{id}:
    parameters:
      - $ref: '#/components/parameters/Id'
    delete:
      tags: [Maquinas]
      summary: Faz exclusao logica de uma maquina
      responses:
        '204': { description: Maquina excluida }
        '401': { $ref: '#/components/responses/UnauthorizedError' }
        '404': { $ref: '#/components/responses/NotFoundError' }
  /rentals:
    get:
      tags: [Alugueis]
      summary: Lista alugueis do usuario
      responses:
        '200':
          description: Lista de alugueis
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Rental' }
        '401': { $ref: '#/components/responses/UnauthorizedError' }
    post:
      tags: [Alugueis]
      summary: Reserva uma maquina
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RentalInput' }
            example:
              machineId: machine-id
              tenant: Joao da Silva
              start: '2026-09-10'
              end: '2026-09-10'
              notes: Pagamento na entrega
      responses:
        '201':
          description: Aluguel criado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Rental' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/UnauthorizedError' }
        '404': { $ref: '#/components/responses/NotFoundError' }
        '409': { $ref: '#/components/responses/ConflictError' }
  /rentals/{id}/finish:
    parameters:
      - $ref: '#/components/parameters/Id'
    post:
      tags: [Alugueis]
      summary: Finaliza um aluguel
      responses:
        '200':
          description: Aluguel finalizado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Rental' }
        '401': { $ref: '#/components/responses/UnauthorizedError' }
        '404': { $ref: '#/components/responses/NotFoundError' }
        '409': { $ref: '#/components/responses/ConflictError' }
  /dashboard:
    get:
      tags: [Dashboard]
      summary: Retorna indicadores do usuario
      responses:
        '200':
          description: Dashboard
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Dashboard' }
        '401': { $ref: '#/components/responses/UnauthorizedError' }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque-token
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema: { type: string }
  responses:
    ValidationError:
      description: Entrada invalida
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          example: { error: Dados de entrada invalidos., code: VALIDATION_ERROR }
    UnauthorizedError:
      description: Sessao invalida
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          example: { error: Sessao invalida., code: UNAUTHORIZED }
    NotFoundError:
      description: Recurso inexistente
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          example: { error: Recurso nao encontrado., code: NOT_FOUND }
    ConflictError:
      description: Conflito
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          example: { error: Recurso em conflito., code: CONFLICT }
  schemas:
    User:
      type: object
      required: [id, nome, email]
      properties:
        id: { type: string, example: user-id }
        nome: { type: string, example: Maria Silva }
        email: { type: string, format: email, example: maria@example.com }
        tipoUsuario: { type: string, example: PRODUTOR }
        telefone: { type: string, nullable: true }
        regiao: { type: string, nullable: true, example: Sul }
        createdAt: { type: string, format: date-time }
    Session:
      type: object
      required: [token, user]
      properties:
        token: { type: string, example: opaque-token-hex }
        expiresAt: { type: string, format: date-time }
        user: { $ref: '#/components/schemas/User' }
    RegisterRequest:
      type: object
      required: [nome, email, senha]
      additionalProperties: false
      properties:
        nome: { type: string, minLength: 1 }
        email: { type: string, format: email }
        senha: { type: string, minLength: 8 }
        regiao: { type: string }
    LoginRequest:
      type: object
      required: [email, senha]
      additionalProperties: false
      properties:
        email: { type: string, format: email }
        senha: { type: string, minLength: 8 }
    ProfileUpdateRequest:
      type: object
      additionalProperties: false
      properties:
        nome: { type: string }
        regiao: { type: string, nullable: true }
    Product:
      type: object
      required: [id, nome, quantidade, unidade, cultivo, regiao, preco, createdAt]
      properties:
        id: { type: string }
        nome: { type: string }
        quantidade: { type: number, format: double, example: 12.5 }
        unidade: { type: string, example: saca }
        cultivo: { type: string, example: organico }
        regiao: { type: string }
        preco: { type: number, format: double, example: 75.5 }
        imagem: { type: string, format: uri, nullable: true }
        createdAt: { type: string, format: date-time }
    ProductInput:
      type: object
      required: [nome, quantidade, unidade, cultivo, regiao, preco]
      additionalProperties: false
      properties:
        nome: { type: string }
        quantidade: { type: number, exclusiveMinimum: 0 }
        unidade: { type: string }
        cultivo: { type: string }
        regiao: { type: string }
        preco: { type: number, exclusiveMinimum: 0 }
        imagem: { type: string, format: uri }
    ProductUpdate:
      allOf:
        - $ref: '#/components/schemas/ProductInput'
      required: []
    Cost:
      type: object
      properties:
        id: { type: string }
        categoria: { type: string, enum: [INSUMO, MAO_DE_OBRA, TRANSPORTE, MANUTENCAO, OUTRO] }
        descricao: { type: string }
        valor: { type: number, format: double }
        data: { type: string, format: date }
        produtoId: { type: string, nullable: true }
    CostInput:
      type: object
      required: [categoria, descricao, valor, data]
      additionalProperties: false
      properties:
        categoria: { type: string, enum: [INSUMO, MAO_DE_OBRA, TRANSPORTE, MANUTENCAO, OUTRO] }
        descricao: { type: string }
        valor: { type: number, exclusiveMinimum: 0 }
        data: { type: string, format: date }
        produtoId: { type: string }
    PricingSimulation:
      type: object
      properties:
        id: { type: string }
        produtoId: { type: string, nullable: true }
        insumos: { type: number }
        maoDeObra: { type: number }
        transporteBase: { type: number }
        manutencao: { type: number }
        distanciaKm: { type: number }
        custoPorKm: { type: number }
        quantidade: { type: number }
        margem: { type: number }
        oferta: { type: number }
        demanda: { type: number }
        custoTotal: { type: number }
        custoUnitario: { type: number }
        precoMinimo: { type: number }
        precoIdeal: { type: number }
        createdAt: { type: string, format: date-time }
    PricingSimulationInput:
      type: object
      required: [insumos, maoDeObra, transporteBase, manutencao, distanciaKm, custoPorKm, quantidade, margem, oferta, demanda]
      additionalProperties: false
      properties:
        insumos: { type: number, exclusiveMinimum: 0 }
        maoDeObra: { type: number, exclusiveMinimum: 0 }
        transporteBase: { type: number, exclusiveMinimum: 0 }
        manutencao: { type: number, exclusiveMinimum: 0 }
        distanciaKm: { type: number, exclusiveMinimum: 0 }
        custoPorKm: { type: number, exclusiveMinimum: 0 }
        quantidade: { type: number, exclusiveMinimum: 0 }
        margem: { type: number, exclusiveMinimum: 0 }
        oferta: { type: number, exclusiveMinimum: 0, maximum: 2 }
        demanda: { type: number, exclusiveMinimum: 0, maximum: 2 }
        produtoId: { type: string }
    Machine:
      type: object
      properties:
        id: { type: string }
        nome: { type: string }
        preco: { type: number }
        tipo: { type: string }
        condicoes: { type: string }
        dias: { type: number }
        imagem: { type: string, format: uri, nullable: true }
        disponibilidade: { type: boolean }
        createdAt: { type: string, format: date-time }
    MachineInput:
      type: object
      required: [nome, preco, tipo, condicoes, dias]
      additionalProperties: false
      properties:
        nome: { type: string }
        preco: { type: number, exclusiveMinimum: 0 }
        tipo: { type: string }
        condicoes: { type: string }
        dias: { type: number, exclusiveMinimum: 0 }
        imagem: { type: string, format: uri }
    Rental:
      type: object
      properties:
        id: { type: string }
        tenant: { type: string }
        machineId: { type: string }
        machineName: { type: string }
        start: { type: string, format: date-time }
        end: { type: string, format: date-time }
        notes: { type: string, nullable: true }
        dailyPrice: { type: number }
        total: { type: number }
        finished: { type: boolean }
        createdAt: { type: string, format: date-time }
        finishedAt: { type: string, format: date-time, nullable: true }
    RentalInput:
      type: object
      required: [machineId, tenant, start, end]
      additionalProperties: false
      properties:
        machineId: { type: string }
        tenant: { type: string }
        start: { type: string, format: date, example: '2026-09-10' }
        end: { type: string, format: date, example: '2026-09-10' }
        notes: { type: string }
    Dashboard:
      type: object
      properties:
        profile: { $ref: '#/components/schemas/User' }
        productsCount: { type: integer }
        machinesCount: { type: integer }
        averageIdealPrice: { type: number }
        costComposition:
          type: array
          items:
            type: object
            properties:
              categoria: { type: string }
              total: { type: number }
        latestSimulationComparison:
          type: object
          properties:
            precoMinimo: { type: number }
            precoIdeal: { type: number }
            diferenca: { type: number }
            percentual: { type: number }
    ErrorResponse:
      type: object
      required: [error, code]
      properties:
        error: { type: string, example: Dados de entrada invalidos. }
        code: { type: string, example: VALIDATION_ERROR }